feat: Add Cdo function and remap in QuickFix
This commit introduces a new function `Cdo` in `my_funcs.lua` that
performs a global substitution in all files listed in the QuickFix list.
The function is then mapped to the `<leader>rw` key combination in the
QuickFix window. This allows users to easily replace a word in all files
listed in the QuickFix list by simply pressing `<leader>rw`.