feat: Add SQL support to null-ls
This commit extends the null-ls configuration to include SQL. It adds
SQL to the list of languages supported by the null-ls server in the
lsp_zero.lua file. It also adds SQL diagnostics and formatting
capabilities to the null_ls.lua file, specifically using the sqlfluff
tool with the TSQL dialect.